refactor(executor): extract expansion logic into separate module
- Create src/executor/ directory structure
- Extract expansion functions to src/executor/expansion.rs:
  * expand_variables_in_string() - handles variable, command substitution, and arithmetic expansion
  * expand_variables_in_args() - processes argument lists
  * expand_wildcards() - handles glob pattern expansion
- Move main executor logic to src/executor/mod.rs
- Make execute_and_capture_output() pub(crate) for use by expansion module
- Add proper module documentation and re-exports
- All 668 tests pass successfully

This is the first step in refactoring the 4,996-line executor.rs into focused,
maintainable modules as outlined in the architecture plan.
